+++ hydra-node/testlib/Test/Hydra/HeadLogic/State.hs
@@ -11,6 +11,7 @@ import Hydra.HeadLogic.State (
   ClosedState (..),
   CoordinatedHeadState (..),
   FanoutMode (..),
+  FanoutStepLanded (..),
   HeadState (..),
   IdleState (..),
   OpenState (..),
@@ -63,6 +64,9 @@ instance (ArbitraryIsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (ClosedS
       <*> arbitrary
 
 instance ArbitraryIsTx tx => Arbitrary (FanoutMode tx) where
+  arbitrary = genericArbitrary
+
+instance ArbitraryIsTx tx => Arbitrary (FanoutStepLanded tx) where
   arbitrary = genericArbitrary
 
 instance (ArbitraryIsTx tx, Arbitrary (ChainStateType tx)) => Arbitrary (PartialFanoutState tx) where